requirements.txt
fastapi
pydantic==1.10.9
sqlalchemy
uvicorn[standard]
PyJWT
python-decouple
psycopg2
passlib
bcrypt
firebase-admin
APScheduler
fastapi-mail